# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: مكتب مصطفى حمزة - Accounting Office API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Expense" }
    post:
      operationId: createExpense
      tags: [expenses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateExpenseInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Expense" }
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: OK
...
  /commands:
    get:
      operationId: listCommands
      tags: [commands]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Command" }
    post:
      operationId: createCommand
      tags: [commands]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCommandInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Command" }
  /commands/{id}:
    delete:
      operationId: deleteCommand
      tags: [commands]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: OK
  /commands/{id}/complete:
    patch:
      operationId: toggleCommandComplete
      tags: [commands]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToggleCommandCompleteInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Command" }
  /commands/{id}/flag:
    patch:
      operationId: toggleCommandFlag
      tags: [commands]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToggleCommandFlagInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Command" }
...
  /declarations/toggle:
    post:
      operationId: toggleDeclaration
      tags: [declarations]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToggleDeclarationInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Declaration" }
components:
  schemas:
...
    EntityType:
      type: string
      enum: [فردي, شركة]
    VatStatus:
      type: string
      enum: [نعم, لا, ربع سنوي]
...
    ExpenseCategory:
      type: string
      enum: [رواتب, ايجار مكتب, كهرباء, انترنت, مصاريف تشغيلية]
    Expense:
      type: object
      additionalProperties: false
      required: [id, item, amount, category, createdAt]
      properties:
        id: { type: integer }
        item: { type: string }
        amount: { type: number }
        category: { $ref: "#/components/schemas/ExpenseCategory" }
        clientId: { type: integer, nullable: true }
        clientName: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        addedById: { type: integer, nullable: true }
        addedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateExpenseInput:
      type: object
      additionalProperties: false
      required: [item, amount, category]
      properties:
        item: { type: string, minLength: 1 }
        amount: { type: number, minimum: 0 }
        category: { $ref: "#/components/schemas/ExpenseCategory" }
        clientId: { type: integer, nullable: true }
        notes: { type: string, nullable: true }
...
    CommandPriority:
      type: string
      enum: [normal, urgent, info]
    Command:
      type: object
      additionalProperties: false
      required: [id, text, priority, target, fromName, createdAt, completed, flagged]
      properties:
        id: { type: integer }
        text: { type: string }
        priority: { $ref: "#/components/schemas/CommandPriority" }
        target: { type: string }
        fromId: { type: integer, nullable: true }
        fromName: { type: string }
        flagged: { type: boolean }
        flaggedAt: { type: string, format: date-time, nullable: true }
        flaggedById: { type: integer, nullable: true }
        flaggedByName: { type: string, nullable: true }
        completed: { type: boolean }
        completedAt: { type: string, format: date-time, nullable: true }
        completedById: { type: integer, nullable: true }
        completedByName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateCommandInput:
      type: object
      additionalProperties: false
      required: [text]
      properties:
        text: { type: string, minLength: 1 }
        priority: { $ref: "#/components/schemas/CommandPriority" }
        target: { type: string }
    ToggleCommandCompleteInput:
      type: object
      additionalProperties: false
      required: [completed]
      properties:
        completed: { type: boolean }
    ToggleCommandFlagInput:
      type: object
      additionalProperties: false
      required: [flagged]
      properties:
        flagged: { type: boolean }
...
    DeclarationType:
      type: string
      enum: [income, vat]
    Declaration:
      type: object
      additionalProperties: false
      required: [clientId, clientName, type, periodKey, periodLabel, dueDate, completed]
      properties:
        clientId: { type: integer }
        clientName: { type: string }
        entityType: { $ref: "#/components/schemas/EntityType", nullable: true }
        vatStatus: { $ref: "#/components/schemas/VatStatus", nullable: true }
        type: { $ref: "#/components/schemas/DeclarationType" }
        periodKey: { type: string }
        periodLabel: { type: string }
        dueDate: { type: string, format: date }
        completed: { type: boolean }
        completedAt: { type: string, format: date-time, nullable: true }
        completedByName: { type: string, nullable: true }
    ToggleDeclarationInput:
      type: object
      additionalProperties: false
      required: [clientId, type, periodKey, completed]
      properties:
        clientId: { type: integer }
        type: { $ref: "#/components/schemas/DeclarationType" }
        periodKey: { type: string }
        completed: { type: boolean }