# lib/api-zod/src/generated/types/createExpenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * مكتب مصطفى حمزة - Accounting Office API
 * OpenAPI spec version: 0.3.0
 */
import type { ExpenseCategory } from "./expenseCategory";

export interface CreateExpenseInput {
  /** @minLength 1 */
  item: string;
  /** @minimum 0 */
  amount: number;
  category: ExpenseCategory;
  clientId?: number | null;
  notes?: string | null;
}