# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * مكتب مصطفى حمزة - Accounting Office API
 * OpenAPI spec version: 0.3.0
 */
import type { ExpenseCategory } from "./expenseCategory";

export interface Expense {
  id: number;
  item: string;
  amount: number;
  category: ExpenseCategory;
  clientId?: number | null;
  clientName?: string | null;
  notes?: string | null;
  addedById?: number | null;
  addedByName?: string | null;
  createdAt: Date;
}