# lib/api-zod/src/generated/types/toggleDeclarationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * مكتب مصطفى حمزة - Accounting Office API
 * OpenAPI spec version: 0.3.0
 */
import type { DeclarationType } from "./declarationType";

export interface ToggleDeclarationInput {
  clientId: number;
  type: DeclarationType;
  periodKey: string;
  completed: boolean;
}